move Snowflake tests
